requirements.txt
discord.py[voice]
google
urbandictionary
chucknorris
xkcd
youtube_dl